Rewrite sort, filter hooks
Don't keep any seperate state for query parameters. Doing so will
make us have to hop through some extra hoops in order to support
things such as reacting to / updating params for the query params
being changed from things like navigation.

Trying to update a seperate state when the query params are changed
will cause an update loop with the past solution.

The new solution is a bit more verbose, but makes sure that the URL
is the single source of truth.