Rework platform structure
Rename the source package to platform, which is in line with the
new documentation.

Rename the k8s package to kubernetes to get rid of "acronyms".

Rework the platform interface to be more losely coupled and to take
advantage of existing open source work to support essentially all
types of image names / references.