Use custom image reference type
The distribution.Reference type was a hassle to mashal to and from
JSON (which was used in the cache, for example). In order to make
it more clear, less magical and more stable, let's just support the
use case we have - images with a tag or digest. Continue to use
the distribution package's parsing as it's more or less the source
of truth of the format.