Rewrite platforms to build a graph
Rewrite platforms to build a well-defined graph, rather than just a
list of images and lists of their parents. This will make it easier
to merge results and provide a nice graph through the APIs rather
than resorting to a lot of duplicated paths.

The graph is a forest, where each root corresponds to a unique OCI
image reference. Other nodes describe how this image is used.