Rewrite cache, implement HTTP cache
Rewrite the cache to only handle bytes, with ability to get the
modification time - must like with a file system. This removes the
unnecessarily complex JSON marshalling that was a bit error prone.

Add a HTTP client with optional caching (a DoCached function), that
uses the new cache interface to cache entire responses. That way
marshalling is always the caller's responsibility and it's well
defined.