Rewrite persistence layer to use sqlite
Use sqlite for persiting data, which allows us to handle queries,
versioning, backups and journaling a lot easier.

Use the new layer in the API, adding support for ordering.

For now, filtering by tags is not supported. Support for sorting by
version was removed as the use case was unclear.