Start rework of database, keep raw image date
Keep track of "raw image data" as identified by a platform. That
way, the worker can later on process raw images and populate the
other tables with actual data. The platform owns the raw image data
and ensures that the table is up-to-date and that there are no
stale references. Upon deletion of a reference, it cascades to all
of the processed tables.