Add GHCR description, package parsing
Add further functionality to the GHCR registry. Fur now, disable
the GitHub workflow as it's unclear what parts make sense to have
in the GHCR part and what parts to have in the GitHub part.

I'm yet to see an image with source annotations, so currently the
gain of having a separate GitHub step seems theoretical.