Redesign dashboard for increased responsiveness
Redesign the dashboard / start page to increase the possible ways we can
improve the layout progressively. The table design required a lot of
space horizontally and was difficult to shrink. With a new card and list
layout, the cards extend vertically and its subcomponents can be shrunk
without much effort.