Add support for watching / polling platforms
Start watching Kubernetes resources for changes, automatically reacting
whenever there's a change as opposed to just check once during startup
or polling for resources. Every 30 minutes (for now), a full resync is
performed, making sure Cupdate's data is up-to-date.

Start polling Docker for changes, using the same interval as is set for
processing. This is done as Docker doesn't have an as easy / natural way
to react on changes as Kubernetes does.