Only fetch tags from OCI registries
Greatly improve the accuracy of identified tags by fetching all tags
that are available from the OCI registry. This comes at a cost of more
(sometimes a great deal more) requests. But as it is a specified part of
the OCI distribution APIs, it should not be of much concern. Also, some
registries (like Docker) respond much faster when no pagination is
requested, potentially due to some aggressive caching on their end.