Schedule inserted images for processing
When images are first added, schedule them for processing. This makes
use of the new rate limited queue to make sure that starting Cupdate
with an empty state does not result in too many requests. It also
ensures that scheduling is optimistic, meaning processing won't be
scheduled if there queue is full.