Use biome instead of eslint and prettier
With eslint there was a lot of config and dependencies to maintain.
Every now and then there were breaking changes taking time to repair.

Biome packs all the features we used in a single tool, hopefully
reducing the maintenance burden. As it's also faster and includes
formatting capabilities, we can use it instead of prettier as well.

There are some few features it does not support, which I'm fine with at
this point. For example, there's no way to specify forbidden syntax. In
the few cases where that was an issue, a review will do for now.

For now, markdown support is only planned, which means that for markdown
prettier is still required. The typical workflow has already been just
running npx prettier -w *.md, so not much will change on that front.