Fix tag selection on small screens
When no images are added (or all are filtered out), the tag selection
would be cut off due to the fix in 8f7278f. The root cause has now been
found to be the tooltips which have an absolute component hovering above
the tooltip's "parent".

Work around the root cause by only rendering the tooltip when hovered,
rather than controlling the visibility.

A proper solution would require additional features not yet available in
all browsers. See #23.