Fix crash when pods are created by unsupported CRD
When Cupdate find a running pod it tries to identify its parent
resource, assuming that it's created by one of the watched resouce
types. When CRDs (or v1/Node kinds) are in use, that's not always true.

Fix this issue by introducing an unknown kind used for non-allow-listed
resource types.

Solves: #54