Mitigate nested anchor tags
Nested anchor tags (as used by the badges of the image card) are not
legal. As we still want the image card to be a proper anchor, let's try
our best to simulate anchors for the badges.

This is not by any means a perfect solution, but will suffice for most
cases.