Fix tag selection click on Windows
On Windows, blur events are not handles the same was as on macOS, for
example. This causes the tag select component to be closed as soon as
the user clicked on another component, event if it was the tag select
itself.

Fix this by instead checking for clicks outside the bounds of the tag
select component, which works across platforms.

Add support for closing the element using the Escape key as that is a
common hotkey on most platforms.