Improve support for references with IPs, digests
Improve support in the UI for references that consist of a host part
that is an IPv4 or IPv6.

Improve support in the UI for references with a host and port as well as
a tag.

Improve support for digests in references.

Add unit tests to assert support in the backend, fix parsing of
references that have a tag and a digest.