Build the Kubernetes graph from the container up
Build the Kubernetes graph from the container up, allowing us to attempt
to identify what image the underlying container runtime is actually
running, by using Kubernetes' container status whenever possible.

In best case scenario, image references identified will be stored with
whatever tag was resolved by the runtime, as well as the actually digest
resolved by the runtime. This gracefully falls back to the tag used in
the actual spec, trying to keep the digest if possible.

This comes with the caveat of removing support for graphing templated
resources like a job's pod template, which in turn has a container
template and which in turn references the image. For now, this is seen
as an acceptable limitation as the use case for the templates were a bit
unclear.

Another caveat is that since the references are used as ids everywhere,
they will now grow with roughly 70B. This is mostly a problem with the
full text indexing of the references, which will not be intutive if the
digests are included. It's also a slight problem when using the APIs as
the URLs will grow to be quite long. This could be solved in upcoming
commits by allowing the reference without digest to be used as an
identifier for as long as there are no collisions.

For now, parts of the processing will use the tag rather than the digest
and there is still no way to identify new digests of the same. These
shortcomings can be addressed in upcoming commits.

Solves: #90