Don't ignore latest tag when reference has digest
When an OCI reference is canonicalized by Cupdate, it ignores the
"latest" tag, assuming it's implicitly understood unless a tag is
specified. When a digest is identified from the runtime, the digest will
be added to the reference. In these cases we cannot ignore the "latest"
tag as it would make it ambigious if the reference was intended as a
"pinned" reference or if it just has a digest resolved by the runtime.