Improve auth support for lscr.io
lscr.io is a pseudo-registry with multiple other backends. If possible,
it uses ghcr.io, with fallbacks on Docker Hub, Quay and GitLab.

For now, assume ghcr.io, but allow the auth for all of the fallbacks to
be used with lscr.io.

Further work is required to automatically detect what auth method to use
for lscr.io, though.