Compare manifests deeply on non-matching digests
When an OCI image index is updated, it receives a new digest. It isn't
necessarily so that that means that the image in use is updated.

Add logic to deeply compare the old and new manifests to not alert on
such errors.

Add platform-dependant logic, without adding full support just yet as
further config would be required.