Store links, vulnerabilities as blobs
As links and vulnerablities were never used in queries, store them as
JSON blobs instead. This enables us to more easily diff changes made to
the tables as each reference will have at most one entry.

The old tables are removed and the new have a different name (they are
given a v2 suffix). This enables a somewhat graceful update where the
data will be dropped without affecting other data. Eventually the data
will be updated again and inserted into the new tables. For v1, rename
the tables again to remove the suffix.