Implicitly depend on OCI job
As the GitHub job already depends on all of the registry jobs, which in
turn depend on the OCI job, the relationship can be implicit.

This has the upside of the dependency graph being a bit more clear as
the "arrow" between the GitHub job and the OCI job is implicit and not
part of the final graph.