Work around testify assertion of time structs
The time structs contain a pointer to the time zone used. This is an
issue when times are created using the time.Local time zone on a machine
using UTC as, when the times are parsed from JSON, they will use the UTC
time zone rather than the local (which is also UTC).