Add workaround for images lacking some data
In a previous commit, the tables for vulnerabilities and links were
changed to handle opaque blobs as their data was not used in queries.

When doing so, the tables were renamed to ensure that there was a
somewhat graceful upgrade path. Unfortunately, this could cause nil
values to be returned where the UI wasn't expecting them, causing it to
fail.

Fix this by checking if no results were found (even though there always
should be).