Implement proper WWW-Authenticate support for OCI
Add support for the WWW-Authenticate header mandated for OCI registries
to allow for "discovery" of the required authentication. This superseeds
the previous per-registry authentication, which removes a lot of
duplicate code.

This feature also makes sure that self-hosted registries work better.

Further work is required to allow for HTTP-only registries, if that's
something we end up wanting.