Clean up vulnerability warning icon
Align the size of the vulnerability warning icon on image cards with the
size of other icons. Fix the SVG having unwanted margins.

Remove the vulnerability warning icon from the image page as there are
several places on that page where vulnerabilities are mentioned, if any
have been found. In addition, the icon was difficult to get right in
terms of alignment and looks.