Extract API to separate lib
Improve testability and mockability as well as general code cleanliness
by extracting the API parts of the code to separate files.

By separating the actual client / API calling and the react layer we
also lay the ground work for mocking the API, which is useful to provide
a demo mode.