Fix GitHub releases not showing
For Docker Hub images that have been connected to a GitHub repository,
we can identify releases, but due depending on the ghcr job, which can
be skipped, the job was not always run.

Solve this by not depending on any job but the OCI one, which is the one
actually in use these days.

Solves: #248