Improve handling of scorecards for cased repos
Improve support for getting OpenSSF scorecards for repositories with
varying casing. The OpenSSF API is case-sensitive, meaning we must rely
on a case-sensitive representation of the repository URL. This can be
retrieved from the determine_source job.

Doing so does remove support for repositories identified via a GHCR
package's package page. Though the use of it seems low and it seems to
sometimes return the non-canonical (lowercased) representation anyway.

This is partly a workaround for ossf/scorecard-webapp#788.