Debounce Kubernetes events
In some Kubernetes environments, resources are changed very frequently
(as in multiple times every minute). This will cause Cupdate to create a
cheap initial graph of the environment. This can result in additional
API calls to the Kubernetes APIs. These are cached, but the traffic can
be unwanted.

To let users control the load, debounce the events by default and make
the debounce interval configurable.