Add a merge gatekeeper
Add an action that will block until all other actions complete. This is
required as we want some changes to trigger workflows that must complete
before merging a PR. But it does not make sense for all PRs to require
the same set of workflows to complete - which is the only control GitHub
grants us. So we need to have this action as a workaround.