Fix single manifest removed in annotation filter
When getting annotations for a reference with support for multiple
architectures, the client would try to filter out all but the ones that
could match the one actually in use. Not all manifests support the
filters, which could lead the client to filter out all manifests. This
would incorrectly cause no annotations to be found.

Solve this by only applying the filter if there are more than one
candidate manifest. This way, if the image is single-arch, the manifest
will still be identified correctly.