Add support for reading refs from static file
Add support for a "static file platform", reading OCI references from a
file as opposed to automatic discovery.

This enables additional use cases where the automatic support is lacking
or not possible. Additionally, it helps with testing as we can define a
wide range of images to test the full set of Cupdate's features.

Solves: #425