Make osv-scanner a runtime dependency
The osv-scanner has been included as a Go dependency, leading to easy
library use. However, the scanner has a lot of dependencies, which in
turn adds a lot of dependencies to Cupdate.

In order to let us focus on dealing with maintaining our direct
dependencies and follow security updates that we are really affected by,
break out the dependency on osv-scanner to a runtime dependency instead.

The binary is included in the official Docker build, which should be the
only one used any way. If the binary is not found, the job fails but the
other jobs of the workflow will continue.