Split optional modules for faster load times
Split the image page and all of its cards into separate modules to
increase load times for users that don't use all features.

It lowers the required JS with about ~1MB if just viewing the dashboard.
It could also lower the required JS a bit when not all features are
used, such as when there are no SBOMs or other attestations.