Fix collision issue in image update table
Fix a collision issue that occurred when the latest reference was not
unique in the images_updates table by creating a compound primary key of
the new and old references, ensuring the pairs are unique. Additionally,
replace entries on collision.

Add additional tests to replicate and regression test this change.

As a part of this change, the images_updates table is dropped and
recreated.