Add support for attestations from referrers
Add support for finding attestations in referrers, as used by Docker's
hardened images.

Reuse the image manifest type for attestation manifests as well as
they're one the same.

Note that this includes a breaking change, the "type" of an SBOM in the
API will now be the predicate type - not just "spdx". This in order to
make it easier to support other predicate types.